Object-relational mapping

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,37 +1,27 @@
 #!/usr/bin/python3
 
-import sys
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from model_state import Base, State
-
 if __name__ == "__main__":
-    mysql_username = sys.argv[1]
-    mysql_password = sys.argv[2]
-    database_name = sys.argv[3]
+    import sys
+    from sys import argv
+    from sqlalchemy import create_engine
+    from sqlalchemy.orm import sessionmaker
+    from model_state import Base, State
 
-    connection_string = (
-        f'mysql+mysqldb://{mysql_username}:{mysql_password}'
-        f'@localhost/{database_name}'
-    )
-    engine = create_engine(connection_string, pool_pre_ping=True)
+    if len(argv) != 4:
+        print("Usage: {} username password database".format(argv[0]))
+        sys.exit(1)
+
+    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'.format(
+        argv[1], argv[2], argv[3]), pool_pre_ping=True)
+
     Base.metadata.create_all(engine)
-
     Session = sessionmaker(bind=engine)
-    session = Session()
 
-    try:
-        states_to_delete = session.query(State).filter(
-                State.name.like('%a%')).all()
-        for state in states_to_delete:
+    with Session() as session:
+        state_delete = session.query(State).filter(
+                  State.name.like('%a%')).all()
+
+        for state in state_delete:
             session.delete(state)
 
         session.commit()
-        print("Deletion completed successfully.")
-
-    except Exception as e:
-        print(f"Error: {e}")
-        session.rollback()
-
-    finally:
-        session.close()
